llm_server/requirements.txt
websockets
openai
whisper
numpy
soundfile
python-dotenv
google-cloud-texttospeech
openai-whisper